+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Fixed signs crashing when placed by a block placer
 + Fixed Mob Crusher sending null text components
 + Allowed to use any item oredicted with "fertilizer" to be used in the Plant fertilizer
++ Reduced Fluid Crafting Processing time by caching missing recipes
 
 # 1.8.6
 + Maybe fixed Mob Crusher sending chat msg on kill